+++ glue_vispy_viewers/scatter/scatter_viewer.py
@@ -1,3 +1,5 @@
+from glue.core.state import lookup_class_with_patches
+
 from ..common.vispy_data_viewer import BaseVispyViewer
 from .layer_artist import ScatterLayerArtist
 from .layer_style_widget import ScatterLayerStyleWidget
@@ -19,7 +21,7 @@ class VispyScatterViewer(BaseVispyViewer):
 
         self._options_widget._update_attributes_from_data(data)
 
-        layer_artist = ScatterLayerArtist(data, vispy_viewer=self._vispy_widget)
+        layer_artist = ScatterLayerArtist(data, vispy_viewer=self)
         self._update_attributes(layer_artist=layer_artist)
 
         self._layer_artist_container.append(layer_artist)
@@ -34,8 +36,8 @@ class VispyScatterViewer(BaseVispyViewer):
         if subset in self._layer_artist_container:
             return
 
-        layer_artist = ScatterLayerArtist(subset, vispy_viewer=self._vispy_widget)
-        self._update_attributes(layer_artist, layer_artist)
+        layer_artist = ScatterLayerArtist(subset, vispy_viewer=self)
+        self._update_attributes(layer_artist=layer_artist)
 
         self._layer_artist_container.append(layer_artist)
 
@@ -53,3 +55,17 @@ class VispyScatterViewer(BaseVispyViewer):
             artist.set_coordinates(self._options_widget.x_att,
                                    self._options_widget.y_att,
                                    self._options_widget.z_att)
+
+    @classmethod
+    def __setgluestate__(cls, rec, context):
+        viewer = super(VispyScatterViewer, cls).__setgluestate__(rec, context)
+        viewer._update_attributes()
+        return viewer
+
+    def restore_layers(self, layers, context):
+        for l in layers:
+            cls = lookup_class_with_patches(l.pop('_type'))
+            props = dict((k, context.object(v)) for k, v in l.items())
+            layer_artist = cls(props['layer'], vispy_viewer=self)
+            self._layer_artist_container.append(layer_artist)
+            layer_artist.set(**props)